chore: Completed TODO task "move to tx-models"
