+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/common/model/SecurityUtils.java
@@ -21,13 +21,13 @@ package org.eclipse.tractusx.traceability.common.model;
 
 
 import org.apache.commons.text.StringEscapeUtils;
-import org.eclipse.tractusx.traceability.qualitynotification.application.alert.request.StartQualityAlertRequest;
-import org.eclipse.tractusx.traceability.qualitynotification.application.base.request.CloseQualityNotificationRequest;
-import org.eclipse.tractusx.traceability.qualitynotification.application.base.request.StartQualityNotificationRequest;
-import org.eclipse.tractusx.traceability.qualitynotification.application.base.request.UpdateQualityNotificationRequest;
 import org.eclipse.tractusx.traceability.qualitynotification.infrastructure.edc.model.EDCNotification;
 import org.eclipse.tractusx.traceability.qualitynotification.infrastructure.edc.model.EDCNotificationContent;
 import org.eclipse.tractusx.traceability.qualitynotification.infrastructure.edc.model.EDCNotificationHeader;
+import qualitynotification.alert.request.StartQualityAlertRequest;
+import qualitynotification.base.request.CloseQualityNotificationRequest;
+import qualitynotification.base.request.StartQualityNotificationRequest;
+import qualitynotification.base.request.UpdateQualityNotificationRequest;
 
 import java.util.List;
 import java.util.Objects;
