+++ tx-models/src/main/java/qualitynotification/base/request/QualityNotificationStatusRequest.java
@@ -1,7 +1,5 @@
 /********************************************************************************
- * Copyright (c) 2022, 2023 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
- * Copyright (c) 2022, 2023 ZF Friedrichshafen AG
- * Copyright (c) 2022, 2023 Contributors to the Eclipse Foundation
+ * Copyright (c) 2023 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -18,18 +16,13 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
+package qualitynotification.base.request;
 
-package org.eclipse.tractusx.traceability.qualitynotification.application.base.request;
+import io.swagger.annotations.ApiModel;
 
-import io.swagger.annotations.ApiModelProperty;
-import jakarta.validation.constraints.NotNull;
-import lombok.Data;
-// TODO move to tx-models
-@Data
-public class UpdateQualityNotificationRequest {
-    @NotNull(message = "status must be present")
-    @ApiModelProperty(example = "ACKNOWLEDGED")
-    private UpdateQualityNotificationStatusRequest status;
-    @ApiModelProperty(example = "The reason.")
-    private String reason;
+@ApiModel(description = "Describes status for closed action")
+public enum QualityNotificationStatusRequest {
+    CLOSED;
+
+
 }
